rename files + shorted some links + add style.css, data_manager.py, df_tmdb_tool.py
