+++ .streamlit/secrets.toml
@@ -1,0 +1,1 @@
+API_KEY="192f96229b83ef4a1fa6add55531870f"
