+++ style.css
@@ -1,0 +1,151 @@
+:root {
+    --background-color: #121212;
+    --primary-color: #01d277;
+    --secondary-color: #ffffff;
+    --text-color: #e0e0e0;
+    --text-muted: #9e9e9e;
+    --font-family: 'Roboto', sans-serif;
+    --border-radius: 8px;
+    --transition-speed: 0.3s ease-in-out;
+}
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    background-color: var(--background-color);
+    color: var(--text-color);
+    font-family: var(--font-family);
+    line-height: 1.6;
+}
+
+header {
+    background-color: #1c1c1c;
+    padding: 10px 20px;
+    position: fixed;
+    width: 100%;
+    z-index: 10;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.5);
+}
+
+header a {
+    color: var(--secondary-color);
+    text-decoration: none;
+    margin-right: 20px;
+    transition: color var(--transition-speed);
+}
+
+header a:hover {
+    color: var(--primary-color);
+}
+
+header .logo {
+    font-size: 24px;
+    font-weight: bold;
+}
+
+.banner {
+    position: relative;
+    height: 0.02vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: 000000;
+    text-align: center;
+    margin-top: 60px;
+}
+
+.banner h1 {
+    font-size: 48px;
+}
+
+.stApp {
+    background-color: #00050d;
+    color: white;
+}
+
+/* Navigation */
+.nav-container {
+    background-color: rgba(0, 5, 13, 0.9);
+    padding: 1rem;
+    position: fixed;
+    top: 0;
+    width: 100%;
+    z-index: 1000;
+    display: flex;
+    align-items: center;
+}
+.nav-links {
+    display: flex;
+    gap: 2rem;
+    margin-left: 2rem;
+}
+.nav-links a {
+    color: #cccccc;
+    text-decoration: none;
+    font-size: 0.9rem;
+}
+.nav-links a:hover {
+    color: white;
+}
+
+/* Movie Details */
+.movie-title {
+    font-size: 2.5rem;
+    margin-bottom: 1rem;
+}
+.movie-info {
+    font-size: 1rem;
+    margin-bottom: 0.5rem;
+}
+.movie-meta {
+    color: #cccccc;
+    font-size: 0.9rem;
+}
+.movie-card {
+    cursor: pointer;
+    transition: transform 0.2s;
+}
+.movie-card:hover {
+    transform: scale(1.05);
+}
+.movie-poster {
+    width: 50%;
+    border-radius: 4px;
+}
+
+/* Cast Section */
+.circular-image {
+    display: block;
+    margin: 0 auto;
+    border-radius: 50%;
+    width: 140px;
+    height: 140px;
+    object-fit: cover;
+}
+.circular-image:hover {
+    transform: scale(1.05);
+}
+.actor-container {
+    text-align: center;
+    margin-bottom: 30px;
+}
+.actor-name {
+    font-weight: bold;
+    margin-top: 10px;
+}
+.actor-role {
+    font-style: italic;
+    color: gray;
+}
+
+/* Recommendations */
+.recommendations {
+    margin-top: 3rem;
+}
